backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.0.0
flask-login>=0.6.0
flask-bcrypt>=1.0.0
flask-limiter>=3.0.0
flask-wtf>=1.1.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
email-validator>=2.0.0
werkzeug>=2.3.0
gunicorn>=21.0.0
redis>=4.5.0